requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.6b0
coreapi==2.3.3
coverage==5.5
Django==3.2.4
django-allauth==0.44.0
django-cors-headers==3.7.0
django-decouple==2.1
django-extensions==3.1.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
flake8==3.9.2
Pillow==8.2.0
psycopg2-binary==2.9.1
PyJWT==2.1.0
python3-openid==3.2.0
six==1.16.0
sqlparse==0.4.1

